Move v1 upgrade stuff into ShaderGraph.V1Upgrade.cs
